api-gateway/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Flask==3.1.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
